fix: loaded by bfloat16

+++ evaluation/inference/loader.py
@@ -10,7 +10,7 @@ def load_model(model_name: str, use_bnb: bool = False, device_map="cuda"):
     if use_bnb:
         bnb_config = BitsAndBytesConfig(
             load_in_4bit=True,
-            bnb_4bit_compute_dtype=torch.float16,
+            bnb_4bit_compute_dtype=torch.bfloat16,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type="nf4"
         )
@@ -18,14 +18,14 @@ def load_model(model_name: str, use_bnb: bool = False, device_map="cuda"):
             model_name,
             quantization_config=bnb_config,
             device_map=device_map,
-            torch_dtype=torch.float16,
+            torch_dtype=torch.bfloat16,
             trust_remote_code=True
         )
     else:
         model = AutoModelForCausalLM.from_pretrained(
             model_name,
             device_map=device_map,
-            torch_dtype=torch.float16,
+            torch_dtype=torch.bfloat16,
             trust_remote_code=True
         )
